Add optional token processing for stat and statx
ALICE now told us that they actually rely on unauthorized stat in some
cases. Thus requiring authorization on stat and statx doesn't work.

This patch modifies alice-token-1 to let it process an authorization
token on stat and statx if it present. If it is not present, stat and
statx will be allowed without authorization.

This means there is no need for alice-token-2. That plugin is removed.